Conversion to TypeScript

+++ src/Body/InjuryCard.tsx
@@ -1,28 +1,39 @@
-import dateUtils from './dateUtils.js';
+import { dateUtils } from './dateUtils';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faTrashAlt, faPencilAlt } from '@fortawesome/free-solid-svg-icons'
 import { useState } from 'react';
-import './injury.css';
-import api from '../api.js';
+import './injuryCard.css';
+import { api } from '../api';
+import { Injury } from '../types';
 
-const Injury = (props) => {
+export interface InjuryCardProps {
+    injury: Injury;
+    reloadInjuries(): void;
+}
+
+export const InjuryCard = (props: InjuryCardProps) => {
 
     const [isInUpdateMode, setIsInUpdateMode] = useState(false); 
     const [newBodyPart, setNewBodyPart] = useState(props.injury.bodyPart);
     
     const handleDeleteInjury = () => {
-        api.deleteInjury(props.injury.injuryId)
-            .then(() => props.reloadInjuries());
+        if (props.injury.injuryId) {
+            api.deleteInjury(props.injury.injuryId)
+                .then(() => props.reloadInjuries());
+        }
     };
 
     const handleSubmitUpdateInjury = () => {
-        const newInjury = {
+        const newInjury: Injury = {
+            injuryId: props.injury.injuryId,
             bodyPart: newBodyPart,
             bodyDiagramCoordinates: props.injury.bodyDiagramCoordinates,
             created: props.injury.created
         }
-        api.updateInjury(props.injury.injuryId, newInjury)
-            .then(() => props.reloadInjuries());
+        if (props.injury.injuryId) {
+            api.updateInjury(props.injury.injuryId, newInjury)
+                .then(() => props.reloadInjuries());
+        }
         setIsInUpdateMode(false)
     };
 
@@ -46,13 +57,21 @@ const Injury = (props) => {
                 <button className='injury-edit-button' onClick={() => setIsInUpdateMode(false)}>Cancel</button>
             </span>
         </> : null;
+    
+    const dateString = () => {
+        if (props.injury.created !== undefined) {
+            return dateUtils.formatDate(new Date(props.injury.created));
+        }
+
+        return '00:00:00'; // default date string when cannot get date.
+    }
 
     return (
         <div className='injury'>
             <>
                 <p className='injury-header-item injury-header-bold'>{props.injury.bodyPart}</p>
                 <p className='injury-header-item injury-header-grey'> · </p>
-                <p className='injury-header-item injury-header-grey'>{dateUtils.formatDate(new Date(props.injury.created))}</p>
+                <p className='injury-header-item injury-header-grey'>{dateString()}</p>
             </>
             {injuryFields}
             {injuryButtons}
@@ -61,4 +80,4 @@ const Injury = (props) => {
     );
 }
 
-export default Injury;
+export default InjuryCard;
